+++ fbw-a32nx/src/systems/instruments/src/EFB/index.tsx
@@ -46,6 +46,7 @@ render(
           paxSign: true,
           rmpVhfSpacing: true,
           satcom: true,
+          avionicsTestMode: true,
         },
         realism: {
           mcduKeyboard: true,